requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
amqp==5.2.0
    # via kombu
anyascii==0.3.2
    # via wagtail
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.11.2
    # via wagtail
billiard==4.2.0
    # via celery
celery[redis]==5.3.6
    # via -r requirements/requirements.in
celery-progress==0.3
    # via -r requirements/requirements.in
certifi==2023.11.17
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cryptography==42.0.1
    # via
    #   django-anymail
    #   pyjwt
defusedxml==0.7.1
    # via
    #   python3-openid
    #   willow
dj-rest-auth==5.0.2
    # via -r requirements/requirements.in
dj-stripe==2.8.3
    # via -r requirements/requirements.in
django==5.0.4
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   dj-stripe
    #   django-allauth
    #   django-allauth-2fa
    #   django-anymail
    #   django-cors-headers
    #   django-filter
    #   django-hijack
    #   django-modelcluster
    #   django-otp
    #   django-permissionedforms
    #   django-taggit
    #   django-treebeard
    #   django-waffle
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   wagtail
django-allauth==0.60.1
    # via
    #   -r requirements/requirements.in
    #   django-allauth-2fa
django-allauth-2fa==0.11.1
    # via -r requirements/requirements.in
django-anymail[mailgun]==10.2
    # via -r requirements/requirements.in
django-cors-headers==4.3.1
    # via -r requirements/requirements.in
django-environ==0.11.2
    # via -r requirements/requirements.in
django-filter==23.5
    # via wagtail
django-hijack==3.4.5
    # via -r requirements/requirements.in
django-modelcluster==6.2.1
    # via wagtail
django-otp==1.3.0
    # via django-allauth-2fa
django-permissionedforms==0.1
    # via wagtail
django-taggit==4.0.0
    # via wagtail
django-treebeard==4.7
    # via wagtail
django-waffle==4.1.0
    # via -r requirements/requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   wagtail
djangorestframework-simplejwt==5.3.1
    # via -r requirements/requirements.in
docutils==0.20.1
    # via -r requirements/requirements.in
draftjs-exporter==2.1.7
    # via wagtail
drf-spectacular==0.27.1
    # via -r requirements/requirements.in
et-xmlfile==1.1.0
    # via openpyxl
filetype==1.2.0
    # via willow
html5lib==1.1
    # via wagtail
idna==3.6
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
kombu==5.3.5
    # via celery
l18n==2021.3
    # via wagtail
mailchimp3==3.0.21
    # via -r requirements/requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
openpyxl==3.1.2
    # via wagtail
pillow==10.2.0
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==0.14.0
    # via willow
prompt-toolkit==3.0.43
    # via click-repl
psycopg2-binary==2.9.9
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
    #   pyjwt
pynacl==1.5.0
    # via -r requirements/requirements.in
pypng==0.20220715.0
    # via qrcode
python-dateutil==2.8.2
    # via celery
python3-openid==3.2.0
    # via django-allauth
pytz==2023.4
    # via
    #   django-modelcluster
    #   djangorestframework
    #   l18n
pyyaml==6.0.1
    # via drf-spectacular
qrcode==7.4.2
    # via django-allauth-2fa
redis==5.0.1
    # via celery
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   django-allauth
    #   django-anymail
    #   mailchimp3
    #   requests-oauthlib
    #   stripe
    #   wagtail
requests-oauthlib==1.3.1
    # via django-allauth
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
sentry-sdk==1.40.0
    # via -r requirements/requirements.in
six==1.16.0
    # via
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
stripe==4.2.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
telepath==0.3.1
    # via wagtail
typing-extensions==4.9.0
    # via qrcode
tzdata==2023.4
    # via celery
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.0.7
    # via
    #   django-anymail
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wagtail==5.2.3
    # via -r requirements/requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
willow[heif]==1.6.3
    # via
    #   wagtail
    #   willow
